requirements.txt
streamlit>=1.32.0
opencv-python>=4.9.0
Pillow>=10.2.0
numpy>=1.26.0
torch>=2.2.0
torchvision>=0.17.0
segment-anything==1.0
rembg[cpu]>=2.1.0
scipy>=1.14.0
google-protobuf>=4.25.0
pytest>=7.0.0
